requirements-dev.txt
mock
pyre-check
pytest-aiohttp
pytest-cov
pytest-mock
pytest-sugar
pytest-timeout
pytest
python-dateutil
pylint
